+++ src/client/java/pl/peterwolf/echoesinink/client/render/PrintingPressRenderState.java
@@ -11,6 +11,9 @@ public class PrintingPressRenderState extends BlockEntityRenderState {
 	public Direction facing = Direction.NORTH;
 	public PressPhase phase = PressPhase.INCOMPLETE;
 	public float animProgress;
+	public float inkingProgress;
+	public boolean matrixInked;
+	public boolean metalTypeMatrix;
 	public boolean hasScrew;
 	public boolean hasHandle;
 	public boolean hasPlaten;
@@ -18,6 +21,7 @@ public class PrintingPressRenderState extends BlockEntityRenderState {
 	public ItemStack output = ItemStack.EMPTY;
 	@Nullable public ItemStackRenderState matrixRenderState;
 	@Nullable public ItemStackRenderState inkRenderState;
+	@Nullable public ItemStackRenderState inkLayerRenderState;
 	@Nullable public ItemStackRenderState sheetRenderState;
 	@Nullable public ItemStackRenderState screwRenderState;
 	@Nullable public ItemStackRenderState handleRenderState;